fix code style according Ashley
